Group all tokenizers under a single module and configure upfront
