--- lib/bumblebee/text/albert_tokenizer.ex
@@ -1,19 +1,0 @@
-defmodule Bumblebee.Text.AlbertTokenizer do
-  @moduledoc """
-  ALBERT tokenizer.
-  """
-
-  import Bumblebee.Shared
-
-  tokenizer_impl(
-    special_tokens: %{
-      bos: "[CLS]",
-      eos: "[SEP]",
-      unk: "<unk>",
-      sep: "[SEP]",
-      pad: "<pad>",
-      cls: "[CLS]",
-      mask: "[MASK]"
-    }
-  )
-end
