--- lib/bumblebee/text/bart_tokenizer.ex
@@ -1,19 +1,0 @@
-defmodule Bumblebee.Text.BartTokenizer do
-  @moduledoc """
-  BART tokenizer.
-  """
-
-  import Bumblebee.Shared
-
-  tokenizer_impl(
-    special_tokens: %{
-      bos: "<s>",
-      eos: "</s>",
-      unk: "<unk>",
-      sep: "</s>",
-      pad: "<pad>",
-      cls: "<s>",
-      mask: "<mask>"
-    }
-  )
-end
