--- lib/bumblebee/text/blenderbot_tokenizer.ex
@@ -1,19 +1,0 @@
-defmodule Bumblebee.Text.BlenderbotTokenizer do
-  @moduledoc """
-  Blenderbot tokenizer.
-  """
-
-  import Bumblebee.Shared
-
-  tokenizer_impl(
-    special_tokens: %{
-      unk: "<unk>",
-      bos: "<s>",
-      eos: "</s>",
-      pad: "<pad>",
-      sep: "</s>",
-      cls: "<s>",
-      mask: "<mask>"
-    }
-  )
-end
