--- lib/bumblebee/text/camembert_tokenizer.ex
@@ -1,19 +1,0 @@
-defmodule Bumblebee.Text.CamembertTokenizer do
-  @moduledoc """
-  Camembert tokenizer.
-  """
-
-  import Bumblebee.Shared
-
-  tokenizer_impl(
-    special_tokens: %{
-      bos: "<s>",
-      eos: "</s>",
-      unk: "<unk>",
-      sep: "</s>",
-      pad: "<pad>",
-      cls: "<s>",
-      mask: "<mask>"
-    }
-  )
-end
