+++ lib/bumblebee/text/text_classification.ex
@@ -31,6 +31,9 @@ defmodule Bumblebee.Text.TextClassification do
     batch_size = compile[:batch_size]
     sequence_length = compile[:sequence_length]
 
+    tokenizer =
+      Bumblebee.configure(tokenizer, length: sequence_length, return_token_type_ids: false)
+
     {_init_fun, predict_fun} = Axon.build(model)
 
     scores_fun = fn params, input ->
@@ -73,10 +76,7 @@ defmodule Bumblebee.Text.TextClassification do
 
       inputs =
         Nx.with_default_backend(Nx.BinaryBackend, fn ->
-          Bumblebee.apply_tokenizer(tokenizer, texts,
-            length: sequence_length,
-            return_token_type_ids: false
-          )
+          Bumblebee.apply_tokenizer(tokenizer, texts)
         end)
 
       batch_key = Shared.sequence_batch_key_for_inputs(inputs, sequence_length)
