+++ lib/bumblebee/text/token_classification.ex
@@ -34,6 +34,13 @@ defmodule Bumblebee.Text.TokenClassification do
     batch_size = compile[:batch_size]
     sequence_length = compile[:sequence_length]
 
+    tokenizer =
+      Bumblebee.configure(tokenizer,
+        length: sequence_length,
+        return_special_tokens_mask: true,
+        return_offsets: true
+      )
+
     {_init_fun, predict_fun} = Axon.build(model)
 
     scores_fun = fn params, input ->
@@ -73,11 +80,7 @@ defmodule Bumblebee.Text.TokenClassification do
 
       all_inputs =
         Nx.with_default_backend(Nx.BinaryBackend, fn ->
-          Bumblebee.apply_tokenizer(tokenizer, texts,
-            length: sequence_length,
-            return_special_tokens_mask: true,
-            return_offsets: true
-          )
+          Bumblebee.apply_tokenizer(tokenizer, texts)
         end)
 
       inputs = Map.take(all_inputs, ["input_ids", "attention_mask"])
