+++ lib/bumblebee/tokenizer.ex
@@ -42,7 +42,7 @@ defmodule Bumblebee.Tokenizer do
   @doc """
   Performs tokenization and encoding on the given input.
   """
-  @callback apply(t(), input() | list(input()), keyword()) :: any()
+  @callback apply(t(), input() | list(input())) :: any()
 
   @doc """
   Decodes a list of token ids into a sentence.
