+++ mix.exs
@@ -104,22 +104,7 @@ defmodule Bumblebee.MixProject do
         ],
         Preprocessors: [
           Bumblebee.Audio.WhisperFeaturizer,
-          Bumblebee.Text.AlbertTokenizer,
-          Bumblebee.Text.BartTokenizer,
-          Bumblebee.Text.BertTokenizer,
-          Bumblebee.Text.BlenderbotTokenizer,
-          Bumblebee.Text.CamembertTokenizer,
-          Bumblebee.Text.ClipTokenizer,
-          Bumblebee.Text.DistilbertTokenizer,
-          Bumblebee.Text.Gpt2Tokenizer,
-          Bumblebee.Text.GptNeoXTokenizer,
-          Bumblebee.Text.LayoutLmTokenizer,
-          Bumblebee.Text.LlamaTokenizer,
-          Bumblebee.Text.MbartTokenizer,
-          Bumblebee.Text.RobertaTokenizer,
-          Bumblebee.Text.T5Tokenizer,
-          Bumblebee.Text.WhisperTokenizer,
-          Bumblebee.Text.XlmRobertaTokenizer,
+          Bumblebee.Text.PreTrainedTokenizer,
           Bumblebee.Vision.BlipFeaturizer,
           Bumblebee.Vision.ClipFeaturizer,
           Bumblebee.Vision.ConvNextFeaturizer,
