--- test/bumblebee/text/roberta_tokenizer_test.exs
@@ -1,41 +1,0 @@
-defmodule Bumblebee.Text.RobertaTokenizerTest do
-  use ExUnit.Case, async: true
-
-  import Bumblebee.TestHelpers
-
-  test "encodes text" do
-    assert {:ok, tokenizer} = Bumblebee.load_tokenizer({:hf, "roberta-base"})
-
-    assert %Bumblebee.Text.RobertaTokenizer{} = tokenizer
-
-    inputs =
-      Bumblebee.apply_tokenizer(tokenizer, [
-        "Test sentence with <mask>.",
-        {"Question?", "Answer"}
-      ])
-
-    assert_equal(
-      inputs["input_ids"],
-      Nx.tensor([
-        [0, 34603, 3645, 19, 50264, 4, 2],
-        [0, 45641, 116, 2, 2, 33683, 2]
-      ])
-    )
-
-    assert_equal(
-      inputs["attention_mask"],
-      Nx.tensor([
-        [1, 1, 1, 1, 1, 1, 1],
-        [1, 1, 1, 1, 1, 1, 1]
-      ])
-    )
-
-    assert_equal(
-      inputs["token_type_ids"],
-      Nx.tensor([
-        [0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0]
-      ])
-    )
-  end
-end
